% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{add_lengthgroups}
\alias{add_lengthgroups}
\title{Add length groups to Gadget StockStdPrinter output}
\usage{
add_lengthgroups(stock_data, length_groups, variable = "number",
  keep_zero_counts = FALSE)
}
\arguments{
\item{stock_data}{A \code{data.frame} of stock data retrieved via
\code{\link{get_stock_std}} or \code{\link{read_gadget_stock_std}}}

\item{length_groups}{Numeric vector of length groups to distribute by.
The difference in values probably should be the same as dl in Gadget model}

\item{variable}{Character vector of the variable on which to distribute
length groups. Default is "number". Can be "number.consumed" or
"biomass.consumed"

Returns a wide \code{data.frame} similar to that of \code{stock_data}, but
with values distributed across each length group which are represented as
each column. The output of \code{add_lengthgroup} is meant to go directly to
\code{\link{survey_gadget}}}

\item{keep_zero_counts}{Logical. Keep year/step/area/age combinations with
no individuals}
}
\description{
Add length groups to Gadget StockStdPrinter output
}
\examples{
cod_lendist <- add_lengthgroups(cod_stock_std$cod0, lengrps)
}
