% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{add_survey_fleet}
\alias{add_survey_fleet}
\title{Simple function to add survey fleets to existing fleet}
\usage{
add_survey_fleet(model_fleet, survey_fleet)
}
\arguments{
\item{model_fleet}{An object of class \code{gadget_fleets} or
\code{gadget_fleet}, or, alternatively, of class \code{gadget_model}}

\item{survey_fleet}{A list of class \code{gadget_fleets}}
}
\value{
The same object as \code{model_fleet}, but updated to include
\code{survey_fleet}
}
\description{
Adds any survey fleets to the existing fleet in a Gadget model
}
\examples{
path <- system.file(gad_mod_dir, package = "gadgetSim")
mod <- read_gadget_model(path = path)
survey_fleet <-
  make_gadget_fleet(spr = list(type = "totalfleet",
                               suitability = constant_suit_formula("spr",
                                                                   "cod",
                                                                   list(1)),
                               amount = data.frame(year = 1,
                                                   step = 1,
                                                   area = 1,
                                                   fleetname = "spr",
                                                   amount = 1)))
mod <- add_survey_fleet(mod, survey_fleet)
}
