% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_quoted_}
\alias{as_quoted_}
\title{Convert input to quoted variables}
\usage{
as_quoted_(x, env = parent.frame())
}
\arguments{
\item{x}{input to quote}

\item{env}{environment in which unbounded symbols in expression should be
evaluated. Defaults to the environment in which \code{as.quoted} was executed}
}
\value{
a list of quoted variables
}
\description{
This function is stolen from plyr::as.quoted. Converts characters to quoted
variables
}
\examples{
as.quoted(c("a", "b", "log(d)"))
}
