% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{call_gadget}
\alias{call_gadget}
\title{Call Gadget executable on the command line}
\usage{
call_gadget(path, switches = list(s = TRUE, i = "params.in"),
  fit.dir = NULL, gadget.exe = "gadget", print.out = TRUE,
  print.err = TRUE)
}
\arguments{
\item{path}{A path to the directory where Gadget files are housed}

\item{switches}{List of named switches to supply to Gadget executable.
Names correspond to the switch; values correspond to additional arguments.
If no argument is needed the value should be logical \code{TRUE}.}

\item{fit.dir}{Directory where final optimized values are supplied. e.g. "WGTS"}

\item{gadget.exe}{The Gadget executable to use}

\item{print.out}{Logical. Should Gadget command line output be printed.}

\item{print.err}{Logical. Should Gadget command line errors be printed.}
}
\value{
If \code{print.out = TRUE} or \code{print.err = TRUE}, a character vector giving the output
of the command, one line per string, else nothing
}
\description{
This function will call the gadget command with the given switches as arguments
}
\examples{
call_gadget("gadget_model", switches = list(s = TRUE, i = "params.in", log = "logfile"))
call_gadget("gadget_model", switches = list(s = TRUE, i = "params.in", main = "main.final"),
            fit.dir = "WGTS")
}
