% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_path}
\alias{check_path}
\title{Paste path variable and file name together if path exists}
\usage{
check_path(x, env = parent.frame())
}
\arguments{
\item{x}{Character. Name of file}
}
\value{
Character. Either the file name or the \code{path} and file name pasted if
\code{path} exists
}
\description{
Paste path variable and file name together if path exists
}
\examples{
check_path("main")
}
