% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stockfile_functions.R
\name{check_sf_names}
\alias{check_sf_names}
\title{Negates a logical test of the presence of a name in stockfile}
\usage{
check_sf_names(name, sf = get("stockfile", env = parent.frame()))
}
\arguments{
\item{name}{Character. The name to be checked}

\item{sf}{Named object. Defaults to stockfile as this is used throughout the
\code{\link{check_stockfile_funs}}}
}
\value{
Logical. TRUE if name is not present. FALSE if it is
}
\description{
This function checks to see if a named element exists in a list and returns TRUE if it does not.
}
\examples{
stockfile <- list(stockname = "cod", minage = 1, maxage = 10, minlength = 1, maxlength = 100)
check_sf_names("^livesonareas")
check_sf_names("^stockname")
}
