% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stockfile_functions.R
\name{check_stockfile_funs}
\alias{check_stockfile_funs}
\alias{check_stock_refweightfile}
\alias{check_stock_grw_eat_len}
\alias{check_stock_growth}
\alias{check_stock_m}
\alias{check_stock_iseaten}
\alias{check_stock_predator}
\alias{check_stock_initcond}
\alias{check_stock_migration}
\alias{check_stock_maturity}
\alias{check_stock_movement}
\alias{check_stock_renewal}
\alias{check_stock_spawning}
\alias{check_stock_straying}
\title{Check various entries when making a stockfile using make_gadget_stockfile()}
\usage{
check_stock_refweightfile(stockfile)

check_stock_grw_eat_len(stockfile, lenaggfile = get("lenaggfile", env =
  parent.frame()))

check_stock_growth(stockfile)

check_stock_m(stockfile)

check_stock_iseaten(stockfile, lenaggfile = get("lenaggfile", env =
  parent.frame()))

check_stock_predator(stockfile)

check_stock_initcond(stockfile)

check_stock_migration(stockfile)

check_stock_maturity(stockfile)

check_stock_movement(stockfile)

check_stock_renewal(stockfile)

check_stock_spawning(stockfile)

check_stock_straying(stockfile)
}
\arguments{
\item{stockfile}{A list that contains stockfile information such as
\code{gadget_stockfile_default}, see source code for \code{\link{make_gadget_stockfile}}}
}
\value{
A list that is a subset of \code{stockfile} containing the appropriate names for each
category being checked
}
\description{
These are non-exported functions that are only used within \code{\link{make_gadget_stockfile}}.
Each checks a different aspect of the required arguments to a Gadget stockfile and either
returns a default value if appropriate or stops function execution if a necessary piece is
missing and a straightforward default does not exists
}
\examples{
see source code for \\code{\\link{make_gadget_stockfile}}
}
