% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{comp_lab}
\alias{comp_lab}
\title{Some default labels}
\format{An object of class \code{character} of length 1.}
\usage{
comp_lab
}
\description{
These are just basic default labels for some things that are used frequently
throughout functions in this package.
}
\keyword{datasets}
