% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_functions.R
\name{f_msy}
\alias{f_msy}
\alias{f_crash}
\alias{f_0.1}
\alias{f_lim}
\title{Compute various fishing mortality reference point}
\usage{
f_msy(fish_mort, nat_mort, ages, growth_fun = vb, growth_params = NULL,
  lw_rel = lw, lw_params = NULL, rec_fun = bev_holt, rec_params = NULL,
  mat_fun = logistic_selectivity, mat_params = NULL,
  sel_fun = logistic_selectivity, sel_params = NULL, ...)

f_crash(fish_mort, nat_mort, ages, growth_fun = vb, growth_params = NULL,
  rec_fun = bev_holt, rec_params = NULL, lw_rel = lw, lw_params = NULL,
  mat_fun = logistic_selectivity, mat_params = NULL,
  sel_fun = logistic_selectivity, sel_params = NULL)

f_0.1(fish_mort, nat_mort, ages, growth_fun = vb, growth_params = NULL,
  lw_rel = lw, lw_params = NULL, sel_fun = logistic_selectivity,
  sel_params = NULL)

f_lim(fish_mort, limit, nat_mort, ages, growth_fun = vb,
  growth_params = NULL, lw_rel = lw, lw_params = NULL,
  rec_fun = bev_holt, rec_params = NULL, mat_fun = logistic_selectivity,
  mat_params = NULL, sel_fun = logistic_selectivity, sel_params = NULL,
  ...)
}
\arguments{
\item{fish_mort}{Numeric. A vector depicting fishing mortality}

\item{nat_mort}{Numeric. A vector of either length 1 or the same length as \code{ages}}

\item{ages}{Numeric vector representing ages}

\item{growth_fun}{The name of the growth function to be used}

\item{growth_params}{List of named elements corresponding to arguments to \code{growth_fun}}

\item{lw_rel}{Function name determining the length-weight relationship}

\item{lw_params}{List of named elements corresponding to arguments to \code{lw_rel}}

\item{sel_fun}{Function determining the length-based selectivity of the fleet}

\item{sel_params}{List of named elements corresponding to arguments to \code{sel_fun}}

\item{...}{}
}
\value{
A numeric vector of length 1, except for \code{f_lim}, which returns 2 values
}
\description{
These functions will compute a variety of reference points relating to fishing mortality.
Various parameters about the stock must be entered as these use yield-per-recruit analysis
to determing the reference points.
}
\details{
\code{f_msy} will determine the fishing mortality that produces maximum sustained yield,
\code{f_crash} returns the fishing mortality that will make the stock go extinct
(i.e. crash the stock), \code{f_0.1} produces the fishing mortality that is roughly 1/10th of
the slope at the origin of the yield-per-recruit curve. \code{f_lim} will find both of the
fishing mortalities to which \eqn{F_{msy}} is scaled by \code{limit}
}
\examples{
f_msy(fish_mort = seq(0, 2, 0.01), nat_mort = 0.2, ages = 0:20,
      growth_fun = vb, growth_params = list(linf = 125, k = 0.15, t0 = -0.5),
      lw_rel = lw, lw_params = list(alpha = 7e-06, beta = 3.1),
      rec_fun = bev_holt, rec_params = list(mu = 4e8, lambda = 1.7e08),
      mat_fun = logistic_selectivity, mat_params = list(alpha = 0.25, l50 = 30),
      sel_fun = logistic_selectivity, sel_params = list(alpha = 0.25, l50 = 40))
}
