% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_defaults.R
\docType{data}
\name{fleet_defaults}
\alias{fleet_defaults}
\alias{totalfleet}
\alias{numberfleet}
\alias{linearfleet}
\alias{effortfleet}
\alias{quotafleet}
\title{Fleet Defaults}
\format{An object of class \code{gadget_fleet} (inherits from \code{list}) of length 5.}
\usage{
totalfleet

numberfleet

linearfleet

effortfleet

quotafleet
}
\description{
Pre-baked default lists to use for reading in fleets
}
\keyword{datasets}
