% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_functions.R, R/update_printfiles.R
\name{format_printfile}
\alias{format_printfile}
\alias{format_printfile}
\title{Format printfile component for writing to printfile}
\usage{
format_printfile(printfile_comp)

format_printfile(printfile_comp)
}
\arguments{
\item{printfile_comp}{A list of class pertaining to a Gadget printfile component,
see \code{\link{update_printfile}}}

\item{printfile_comp}{A list of class pertaining to a Gadget printfile component,
see \code{\link{update_printfile}}}
}
\description{
Format printfile component for writing to printfile

Format printfile component for writing to printfile
}
