% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_defaults.R
\docType{data}
\name{gadget_defaults}
\alias{gadget_defaults}
\alias{gad_mod_dir}
\alias{gadget_main_default}
\alias{gadget_timefile_default}
\alias{gadget_areafile_default}
\alias{gadget_stockfile_default}
\alias{gadget_tagfile_default}
\alias{gadget_otherfood_default}
\title{Gadget defaults}
\format{An object of class \code{character} of length 1.}
\usage{
gad_mod_dir

gadget_main_default

gadget_timefile_default

gadget_areafile_default

gadget_stockfile_default

gadget_tagfile_default

gadget_otherfood_default
}
\description{
Some pre-baked defaults to use in setting up Gadget simulations
}
\keyword{datasets}
