% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gadget_simulate}
\alias{gadget_simulate}
\alias{add_lengthgroups}
\alias{survey_gadget}
\title{Create length distributions and sample Gadget output}
\usage{
add_lengthgroups(stock_data, length_groups, keep_zero_counts = FALSE)

survey_gadget(stock_data, length_groups, survey_suitability, survey_sigma)
}
\arguments{
\item{stock_data}{A \code{data.frame} of stock data retrieved via \code{\link{read_gadget_stock_std}}}

\item{length_groups}{Numeric vector of length groups to distribute by. Probably should be the
same as dl in Gadget model}

\item{keep_zero_counts}{Logical. Keep year/step/area/age combinations with no individuals}

\item{survey_suitability}{Numeric vector the same length as \code{length_groups} representing the
selection probability for each length in \code{length_groups}}

\item{survey_sigma}{Numeric value of multiplicative error to place on samples}
}
\value{
\code{add_lengthgroup} returns a wide \code{data.frame} similar to that of
\code{stock_data}, but with values distributed across each length group which are represented
as each column. The output of \code{add_lengthgroup} is meant to go directly to
\code{survey_gadget} which returns a \code{data.frame} similar to \code{stock_data}, but
disaggregated by length.
}
\description{
These functions create length distributions from Gadget StockStdPrinter output and sample
that output to simulate surveys. Error can be added to simulated surveys
}
\details{
Length-structured population information from Gadget is output as mean length and standard
deviation for each year, step, area, and age combination. \code{add_lengthgroups} takes
output from \code{\link{get_stock_std}} and distributes the number for each respective
combination into numbers at each length specified by \code{length_groups}.
The return value for \code{add_lengthgroups} is a wide \code{data.frame} that can then be
fed into \code{survey_select}, which simulates surveys given the selectivity provided in
\code{survey_suitability} and error given by \code{survey_sigma}
}
\examples{
path <- system.file(gad_mod_dir, package = "gadgetSim")
cod_stock_std <- get_stock_std(main = "WGTS/main.final", params_file = "WGTS/params.final",
                               path = path, fit_dir = "WGTS")
lengrps <- seq(0.5, 50.5, by = 1)
cod_lendist <- add_lengthgroups(cod_stock_std$cod0, lengrps)
}
