% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gadget_files.R
\name{gadget_sub_dir}
\alias{gadget_sub_dir}
\alias{is_gadget_sub_dir}
\title{Write out a Gadget sub-directory and return the path to it}
\usage{
gadget_sub_dir(sub_dir, path = NULL)

is_gadget_sub_dir(obj)
}
\arguments{
\item{sub_dir}{Character. Path to the sub-directory you wish to create}

\item{path}{Optional. Character string of the path of the Gadget model you
are working in}
}
\value{
A character string of the path to the sub-directory specified in
\code{sub_dir}
}
\description{
This function will create a new directory by the name of \code{sub_dir}
within the directory specified by \code{path}. The address to the new
sub-directory is returned.
}
\details{
\code{gadget_sub_dir} is essentially just an easy way to create a
sub-directory of the chosen Gadget model for writing model files to.
It could be used for writing replicant models files.
}
\examples{
\dontrun{
path <- system.file(gad_mod_dir, package = "gadgetSim")
gadget_sub_dir("test_model", path = path)
}
gadget_sub_dir()
}
