% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_defaults.R
\docType{data}
\name{gadgetDefaults}
\alias{gadgetDefaults}
\alias{gad_mod_dir}
\alias{gadget_main_default}
\title{Gadget defaults}
\format{An object of class \code{character} of length 1.}
\usage{
gad_mod_dir

gadget_main_default
}
\description{
Some pre-baked defaults to use in setting up Gadget simulations
}
\keyword{datasets}
