% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_index}
\alias{get_index}
\title{Get the index of obect(s) in a vector}
\usage{
get_index(x, vec)
}
\arguments{
\item{x}{Object for which the index will be retrieved}

\item{vec}{Vector against which x will be indexed}
}
\value{

}
\description{
Get the index of obect(s) in a vector
}
\examples{
get_index(2, 2:10)
get_index("q", letters)
}
