% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{get_si_data}
\alias{get_si_data}
\title{Strip survey samples for the desired length or age groups to use for
surveyindices data}
\usage{
get_si_data(data_samples, by = "lengths", groups)
}
\arguments{
\item{data_samples}{Data.frame of sampled model output as returned by
\code{\link{survey_gadget}}, for example}

\item{by}{Character. Currently either only "lengths" or "ages"}

\item{groups}{Numeric. The values by which to subset \code{by}}
}
\value{
A data.frame of filtered or subsetted data
}
\description{
Easily pick out the data that you want to use for surveyindices
}
\examples{
dat <- expand.grid(year = 1:10, step = 1, area = 1,
                   ages = 1:10, lengths = 1:100)
dat$number <- sample(1:1000, nrow(dat), replace = TRUE)
si_dat <- get_si_data(dat, by = "lengths", groups = 1:20)
}
