% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_stock_std}
\alias{get_stock_std}
\title{Make and retrieve output from the StockStdPrinter printfile component of a Gadget model}
\usage{
get_stock_std(main = "main", params_file = "params.in", path = NULL,
  fit_dir = NULL, gadget_exe = "gadget", ...)
}
\arguments{
\item{main}{Optional. A list of class \code{gadget.main}}

\item{params_file}{Character. Path to the params file used to call Gadget}

\item{path}{Optional. Character vector of path to the directory where Gadget files are located}

\item{fit_dir}{Directory where final optimized values are supplied. (e.g. "WGTS")}

\item{gadget_exe}{The Gadget executable to use}

\item{...}{Additional arguments to include}
}
\value{
List of \code{data.frame}s, one for each stock, of output from StockStdPrinter
printfile component
}
\description{
Make and retrieve output from the StockStdPrinter printfile component of a Gadget model
}
\examples{
\dontrun{
path <- system.file(gad_mod_dir, package = "gadgetSim")
stocks_data <- get_stock_std(path = path)
}
}
