% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_info.R
\name{getStockInfo}
\alias{getStockInfo}
\alias{get_stock_ages}
\alias{get_stock_areas}
\alias{get_stocknames}
\alias{get_stock_lengths}
\title{Functions to retrieve basic information about Gadget stocks}
\usage{
get_stock_ages(stocks)

get_stock_areas(stocks)

get_stocknames(stocks)

get_stock_lengths(stocks)
}
\arguments{
\item{stocks}{A list of class \code{gadget.stocks} containing one (or typically more)
lists of class \code{gadget.stock}}
}
\value{
A vector of values found in the Gadget stock file
}
\description{
Functions to retrieve basic information about Gadget stocks
}
\examples{
cod <- read_gadget_stockfiles("cod", path = gad_mod_dir)
get_stock_ages(cod)

}
