% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_functions.R
\name{growth_funs}
\alias{growth_funs}
\alias{vb}
\alias{lw}
\title{Basic functions for growth curves}
\usage{
vb(linf, k, t0, age)

lw(alpha, beta, length)
}
\arguments{
\item{linf}{Numeric. Parameter specifying aximum length at age}

\item{k}{Numeric. Parameter specifying steepness of the growth curve}

\item{t0}{Numeric. Shifting parameter that specifies theoretical age at length 0}

\item{age}{Numeric vector of ages}
}
\value{
A numeric vector corresponding to mean length-at-age for \code{vb} or mean
weight-at-length for \code{lw}
}
\description{
\code{vb} produces the von Bertalanffy growth curve for length at age
\deqn{L_{\infty} * (1 - \exp(-k * (a - t_0)))} where \eqn{a} is age. \code{lw} returns the length-weight
relationship as determined by parameters \eqn{\alpha} and \eqn{\beta} in the function
\deqn{\alpha * L^{\beta}}, where \eqn{L} is length.
}
\examples{
age <- 1:20
length <- vb(125, 0.15, -0.5, age)
wt <- lw(length, 1e-06, 3)
}
