% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_list_element_null}
\alias{is_list_element_null}
\title{Check to see if elements of a list are null}
\usage{
is_list_element_null(lst, keep_names = FALSE)
}
\arguments{
\item{lst}{Any list}
}
\value{
A logical vector of length \code{lst}
}
\description{
Check to see if elements of a list are null
}
\details{
\code{is.null} does not work to check individual elements of a list.
This function does that an returns a logical vector the length of \code{lst}
stating whether each element of \code{lst} is null (TRUE) or not (FALSE)
}
\examples{
is_list_element_null(list(a = 1, b = 2, c = NULL, d = NULL))
}
