% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_functions.R
\name{make_aggfiles}
\alias{make_aggfiles}
\title{Make aggregation files for Gadget printfile components that require them}
\usage{
make_aggfiles(printfile_comp, aggfile_dir, path = NULL)
}
\arguments{
\item{printfile_comp}{A list of class pertaining to a Gadget printfile component,
see \code{\link{update_printfile}}}

\item{aggfile_dir}{Character. Path of the directory where aggfiles will be printed}

\item{path}{Character. Path of the Gadget model}
}
\value{
A list the same class as \code{printfile_comp}, but updated with the
appropriate aggregate files as attributes
}
\description{
Make aggregation files for Gadget printfile components that require them
}
\examples{
cod <- list(stocknames = "cod")
cod <- update_printfile(stock, cod)
cod <- make_aggfiles(cod, aggfile_dir = "print.aggfiles", path = gad_mod_dir)
}
