% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_likelihood.R
\name{make_gadget_likelihood}
\alias{make_gadget_likelihood}
\title{Create a Gadget likelihood file}
\usage{
make_gadget_likelihood(..., penalty = TRUE, understocking = TRUE)
}
\arguments{
\item{...}{Various Gadget likelihood components desired that are of class
\code{gadget_*_likelihood}, see \code{\link{gadget_likelihood_comps}}}

\item{penalty}{Logical. Should a penalty likelihood component be added}

\item{understocking}{Logical. Should an understocking likelihood component be
added}
}
\value{
A list of class \code{gadget_likelihood} that contains the various
likelihood components
}
\description{
This function concentrates the various likelihood components created by
\code{\link{gadget_likelihood_comps}} and puts them all in a single object
ready for writing.
}
\examples{
# see examples in ?gadget_likelihood_comps
lik_file <- make_gadget_likelihood(spr_ldist_comp, spr_aldist_comp)
}
