% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{make_gadget_mainfile}
\alias{make_gadget_mainfile}
\title{Create a list of class "gadget_main"}
\usage{
make_gadget_mainfile(...)
}
\arguments{
\item{...}{Either a list with named objects pertaining to the main file for
Gadget or any number of objects named similarly}
}
\value{
A list of class \code{gadget_main}
}
\description{
Create a list of class "gadget_main"
}
\examples{
main <- list(timefile = "time", areafile = "area", stockfiles = "cod")
make_gadget_mainfile(main)
}
