% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{make_gadget_optinfofile}
\alias{make_gadget_optinfofile}
\title{Create a Gadget optimization info file}
\usage{
make_gadget_optinfofile(...)
}
\arguments{
\item{...}{Named list(s) with names corresponding to the optimization type
desired.}
}
\value{
A list of class \code{gadget_optinfofile}
}
\description{
This function creates a list of class \code{gadget_optinfofile} that can be
used to write to file the components desired for optimization in Gadget.
Arguments to the function can either be left blank for the default; otherwise
they must be named arguments with names corresponding to the optimization
routine desired (see Gadget User Guide. Chapter 11.).
}
\details{
If the default values provided by gadgetSim are desired, then use a named
argument with "default" (i.e. simann = "default"). Otherwise, you can specify
as many or as few parameters to alter in a list
}
\examples{
opt <- make_gadget_optinfofile()
opt <- make_gadget_optinfofile(simann = "default")
opt <- make_gadget_optinfofile(simann = list(simanniter = 10000,
                                             t = 3000),
                               bfgs = list(bfgsiter = 20000))
}
