% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{make_gadget_spawnfile}
\alias{make_gadget_spawnfile}
\title{Create a list of class "gadget_spawnfile" that can be used to write to file}
\usage{
make_gadget_spawnfile(stockname, start_year, end_year, ...)
}
\arguments{
\item{stockname}{Character. The stockname desired (used to define switches)}

\item{start_year}{Numeric. The first year spawning will take place}

\item{end_year}{Numeric. The last year spawning will take place}

\item{...}{Other vectors to be added. Should be named elements corresponding to those needed
for a Gadget spawnfile, see Gadget User Guide.}
}
\value{
A list of class "gadget_spawnfile"
}
\description{
This function creates a list of class "gadget_spawnfile". This list can then be used to
write a spawnfile for a stock to be used by Gadget
}
\examples{
make_gadget_spawnfile("cod", 1, 10)
}
