% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{make_gadget_stockfile}
\alias{make_gadget_stockfile}
\title{Make Gadget stockfile}
\usage{
make_gadget_stockfile(...)
}
\arguments{
\item{...}{A list of named elements. The names correspond to arguments found in Gadget stockfiles}
}
\value{
A list of class \code{gadget.stock} containing information about that stock to be used
by Gadget
}
\description{
Make a list of class \code{gadget.stock} that can be used to write out a Gadget stockfile
}
\examples{
stock_info <-
    list(stockname = "cod", livesonareas = 1, minage = 1, maxage = 10,
         minlength = 1, maxlength = 100, dl = 1)
grw_params <-
    c(to_gadget_formula(quote(cod.linf)), to.gadget.formula(quote(cod.k)),
      0.001, 3)
stock_growth <-
    list(doesgrow = 1, growthfunction = "lengthvbsimple",
         growthparameters = grw_params)
stock_m <- list(naturalmortality = rep(0.2, 10))
stock_list <- c(stock_info, stock_growth, stock_m)
make_gadget_stockfile(stock_list)
}
