% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{make_gadget_strayfile}
\alias{make_gadget_strayfile}
\title{Create a list of class \code{gadget_strayfile} to add to a Gadget stockfile}
\usage{
make_gadget_strayfile(steps, areas, stock_ratios, prop_fun)
}
\arguments{
\item{steps}{Numeric. The steps in which the stock will stray}

\item{areas}{Numeric. The areas in which the stock will stray}

\item{stock_ratios}{Data.frame or list of the stocks and straying ratios}

\item{prop_fun}{Character. The proportionfunction to use}
}
\value{
A list of class \code{gadget_strayfile}
}
\description{
This function creates a list of class \code{gadget_strayfile} that can be
added into a \code{gadget_stock}. The strayfile will then be written as an
attribute at the time of writing
}
\examples{
make_gadget_strayfile(steps = 1, areas = 1,
                      stock_ratios = data.frame("cod", 1),
                      prop_fun = c("constant", 1))
}
