% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{make_gadget_timefile}
\alias{make_gadget_timefile}
\title{Make Gadget timefile}
\usage{
make_gadget_timefile(st_year, end_year, timesteps)
}
\arguments{
\item{st_year}{Integer. The starting year of the Gadget model}

\item{end_year}{Integer. The final year of the Gadget model}

\item{timesteps}{Either a numeric vector giving the number of timesteps and the breakdown of
months, or, more commonly and conventiently one of "annually", "biannually", "quarterly",
"monthly'}
}
\value{
A list of class \code{gadget_time}
}
\description{
Make Gadget timefile
}
\examples{
make_gadget_timefile(1985, 2015, "quarterly")
make_gadget_timefile(1985, 2015, c(6, 2, 1, 3, 3, 1, 2))
}
