% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_list_at_index}
\alias{make_list_at_index}
\title{Split a vector into a list based on indices}
\usage{
make_list_at_index(vec, ind, keep_indices = TRUE)
}
\arguments{
\item{vec}{A vector of any length or class}

\item{ind}{Numeric vector depicting the indices at which to split the list}

\item{keep_indices}{Logical. Return the indices in each element of the list or not}
}
\value{
A list with length equal to that of the lenght of \code{vec}. Each element
contains the values in between the indices including the values at the indices if
\code{keep_indices = TRUE}, the default.
}
\description{
This function takes a vector and indices from that vector and splits the vector into a
list at the assigned indices. Values at the indices can either be kept or discarded.
}
\examples{
a <- letters
ind <- grep("a|e|i|o|u", a)
consonants <- make_list_at_index(a, ind, keep_indices = FALSE)
}
