% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{make_switches}
\alias{make_switches}
\title{Turn any character vector into a switch}
\usage{
make_switches(...)
}
\arguments{
\item{...}{Any number of character objects}
}
\value{
A list with length the same number of items entered in \code{...}.
All items will have a "#" placed in front of them
}
\description{
Character vectors are strictly returned when using
\code{\link{to_gadget_formula}}. This function allows one to explicitly
turn a character vector into a switch
}
\examples{
make_switches("linf", "k")
stockname <- "cod"
make_switches(paste(stockname, "linf", sep = "."),
              paste(stockname, "k", sep = "."))
}
