% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_functions.R
\name{makeAggfileHelpers}
\alias{makeAggfileHelpers}
\alias{make_agg_dir}
\alias{make_area_agg_dir}
\alias{make_age_agg_dir}
\alias{make_len_agg_dir}
\alias{make_predlen_agg_dir}
\alias{make_preylen_agg_dir}
\alias{make_area_aggfile}
\alias{make_age_aggfile}
\alias{make_len_aggfile}
\alias{make_predlen_aggfile}
\alias{make_preylen_aggfile}
\alias{check_agg_type}
\alias{isNULL_aggfiles}
\title{Functions to make the area, age, length, predator, and prey aggfiles}
\usage{
make_agg_dir(stocks, aggfile_dir, agg_type)

make_area_agg_dir(stocks, aggfile_dir)

make_age_agg_dir(stocks, aggfile_dir)

make_len_agg_dir(stocks, aggfile_dir)

make_predlen_agg_dir(stocks, aggfile_dir)

make_preylen_agg_dir(stocks, aggfile_dir)

make_area_aggfile(stocks)

make_age_aggfile(stocks)

make_len_aggfile(stocks, len_type = "length")

make_predlen_aggfile(stocks)

make_preylen_aggfile(stocks)

check_agg_type(printfile_comp, agg_type, stocks, aggfile_dir)

isNULL_aggfiles(printfile_comp)
}
\arguments{
\item{stocks}{A list of class \code{gadget_stock} or \code{gadget_stocks}}

\item{aggfile_dir}{Character. Path of the directory where aggfiles will be printed}

\item{printfile_comp}{A list of class pertaining to a Gadget printfile component,
see \code{\link{update_printfile}}}
}
\value{
\code{make_*_agg_dir} returns a character vector of a path to the
printfile location. \code{make_*_aggfile} returns a character vector of the
aggregate file itself ready to be written to file. The \code{check_*} functions
look to see if the appropriate aggregate file is present, and, if not, will
write the aggfile path to the aggfile line in \code{printfile_comp} of
\code{\link{make_aggfiles}} and the aggfile as an attribute to \code{printfile_comp}.
}
\description{
These are helper functions to the function \code{\link{make_aggfiles}}.
\code{make_<type>_agg_dir} functions return a path to the printfile,
\code{make_<type>_aggfile} produces the content for the aggfile itself,
and the \code{check_*} family of functions checks for the presence of an
aggfile in the \code{printfile_comp} of \code{\link{make_aggfiles}} and
returns the aggfile and path if not present.
}
\examples{
main <- read_gadget_main(path = gad_mod_dir)
stocks <- read_gagdget_stockfiles(main = main, path = gad_mod_dir)
make_area_agg_dir(stocks, aggfile_dir = "print.aggfiles")
make_area_aggfile(stocks)
}
