% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_functions.R
\name{makeAggfiles}
\alias{makeAggfiles}
\alias{make_area_agg_dir}
\alias{make_area_aggfile}
\alias{make_age_agg_dir}
\alias{make_age_aggfile}
\alias{make_len_agg_dir}
\alias{make_len_aggfile}
\alias{check_area_agg}
\alias{check_age_agg}
\title{Functions to make the area, age, length, predator, and prey aggfiles}
\usage{
make_area_agg_dir(stocks, aggfiles)

make_area_aggfile(stocks)

make_age_agg_dir(stocks, aggfiles)

make_age_aggfile(stocks)

make_len_agg_dir(stocks, aggfiles)

make_len_aggfile(stocks)

check_area_agg()

check_age_agg()
}
\arguments{
\item{stocks}{A list of class \code{gadget.stock} or \code{gadget.stocks}}

\item{aggfiles}{Character. Path of the directory to print aggfiles}
}
\value{
Character vector ready to be written to file
}
\description{
Functions to make the area, age, length, predator, and prey aggfiles
}
\examples{
main <- read_gadget_main(path = gad_mod_dir)
stocks <- read_gagdget_stockfiles(main = main, path = gad_mod_dir)
make_area_agg_dir(stocks, aggfiles = "print.aggfiles")
make_area_aggfile(stocks)
}
