% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{params}
\alias{params}
\alias{make_gadget_params}
\alias{init_params}
\title{Create a data.frame of class "gadget_params"}
\usage{
make_gadget_params()

init_params(df, switch, value, lower, upper, optimize = TRUE)
}
\value{
A data.frame of class \code{gadget_params}. \code{make_gadget_params}
returns an empty data.frame, and \code{init_params} updates the parameter
found by \code{switch}.
}
\description{
These functions create and update a data.frame of class \code{gadget_params}.
This is the parameter file for switches used in a Gadget model
}
\examples{
params <- make_gadget_params()
params <- init_params(params, "linf", 125, 115, 135)
params <- init_params(params, "k", 0.25, 0.1, 0.3, optimize = FALSE)

# setup for efficient use with dplyr
params <-
  make_gadget_params() \%>\%
  init_params("linf", 125, 115, 135) \%>\%
  init_params("k", 0.25, 0.1, 0.3, optimize = FALSE)
}
