% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_functions.R
\name{printfile_name_check}
\alias{printfile_name_check}
\title{Check the names of a printfile component to ensure all mandatory arguments are present}
\usage{
printfile_name_check(printfile_comp)
}
\arguments{
\item{printfile_comp}{A list of class pertaining to a Gadget printfile component,
see \code{\link{update_printfile}}}
}
\value{
NULL. Only returns an error message if any non-optional arguments are missing
}
\description{
Check the names of a printfile component to ensure all mandatory arguments are present
}
\examples{
printfile_name_check(stock_std)

cod <- update_printfile(stock_std, list(stockname = "cod", printfile = "printfile"))
printfile_name_check(cod)
}
