% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_defaults.R
\docType{data}
\name{printfileArgs}
\alias{printfileArgs}
\alias{stock_std_args}
\alias{stock_full_args}
\alias{stock_args}
\alias{predator_args}
\alias{predator_over_args}
\alias{prey_over_args}
\alias{stock_prey_full_args}
\alias{stock_prey_args}
\alias{predator_prey_args}
\alias{likelihood_args}
\alias{likelihood_summary_args}
\title{Non-optional arguments to Gadget printfile components}
\format{An object of class \code{character} of length 5.}
\usage{
stock_std_args

stock_full_args

stock_args

predator_args

predator_over_args

prey_over_args

stock_prey_full_args

stock_prey_args

predator_prey_args

likelihood_args

likelihood_summary_args
}
\description{
Pre-baked defaults of necessary printfile arguments for use in various functions
}
\keyword{datasets}
