% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_defaults.R
\docType{data}
\name{printfileDefaults}
\alias{printfileDefaults}
\alias{stock_std}
\alias{stock_full}
\alias{stock}
\alias{predator}
\alias{predator_over}
\alias{prey_over}
\alias{stock_prey_full}
\alias{stock_prey}
\alias{predator_prey}
\alias{likelihood}
\alias{likelihoodsummary}
\title{Gadget print file defaults}
\format{An object of class \code{stock_std} (inherits from \code{list}) of length 7.}
\usage{
stock_std

stock_full

stock

predator

predator_over

prey_over

stock_prey_full

stock_prey

predator_prey

likelihood

likelihoodsummary
}
\description{
Pre-baked defaults of printfile types for use in functions
}
\keyword{datasets}
