% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gadget_auxiliaries.R
\name{read_gadget_aux_files}
\alias{read_gadget_aux_files}
\alias{read_gadget_refweightfile}
\alias{read_gadget_stock_len_aggfile}
\alias{read_gadget_preylengths}
\alias{read_gadget_init_cond}
\alias{read_gadget_yearstepfile}
\alias{read_gadget_migration_file}
\alias{read_gadget_maturity_file}
\alias{read_gadget_renewal}
\alias{read_gadget_spawnfile}
\alias{read_gadget_strayfile}
\alias{format_auxiliary_file}
\alias{format_data_dist_file}
\title{Read auxiliary Gadget files (i.e. data files, aggfiles, auxiliary stockfiles,
etc.)}
\usage{
read_gadget_refweightfile(file, path = NULL)

read_gadget_stock_len_aggfile(file, path = NULL)

read_gadget_preylengths(file, path = NULL)

read_gadget_init_cond(file, data_dist_type, path = NULL)

read_gadget_yearstepfile(file, path = NULL)

read_gadget_migration_file(file, migration_type, path = NULL)

read_gadget_maturity_file(file, mat_fun_type, path = NULL)

read_gadget_renewal(file, data_dist_type, path = NULL)

read_gadget_spawnfile(file, path = NULL)

read_gadget_strayfile(file, path = NULL)

format_auxiliary_file(aux_file, aux_names)

format_data_dist_file(data_dist_file, data_names)
}
\arguments{
\item{file}{Character. The name of the file to be read}

\item{path}{Optional. Character vector of path to the directory where Gadget files are located}
}
\value{
Typically a list with the elements specific to what is needed for
each type of auxiliary file. Aggfiles and data files will often return a
data.frame
}
\description{
This suite of functions are specific to types of Gadget auxiliary stockfiles,
with special attention paid to the various auxiliary stockfiles such as
refweightfile, initial condition and renewal data distribution files,
spawnfile, etc. These functions are for use in \code{\link{check_stockfile}}
}
\examples{
gad_data <- read_gadget_refweightfile("Modelfiles/cod.refweightfile",
                                      path = gad_mod_dir)
}
