% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gadget_files.R
\name{read_gadget_fleet}
\alias{read_gadget_fleet}
\title{Read Gadget fleet file}
\usage{
read_gadget_fleet(fleetfiles, main = NULL, path = NULL)
}
\arguments{
\item{fleetfiles}{Character name of fleet file in Gadget model as specified in main file}

\item{main}{Optional. Object of class \code{gadget.main} used to search for fleet file}

\item{path}{Character name of path to Gadget model}
}
\value{
A list detailing the fleets included in Gadget model,
stocks included, and suitability
}
\description{
Read Gadget fleet file
}
\examples{
main <- read_gadget_main(path = system.file(gad_mod_dir, package = "gadgetSim"))
read_gadget_fleet(main = main, path = system.file(gad_mod_dir, package = "gadgetSim"))
}
