% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gadget_files.R
\name{read_gadget_likelihood}
\alias{read_gadget_likelihood}
\title{Read Gadget likelihood file}
\usage{
read_gadget_likelihood(likelihoodfiles, main = NULL, path = NULL)
}
\arguments{
\item{likelihoodfiles}{Character. The name of the likelihood file}

\item{path}{}
}
\description{
Read Gadget likelihood file
}
