% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gadget_files.R
\name{read_gadget_main}
\alias{read_gadget_main}
\title{Read Gadget mainfile}
\usage{
read_gadget_main(file = "main", path = NULL)
}
\arguments{
\item{file}{Character. Name of the main file to be read}

\item{path}{Optional. Character vector of path to the directory where Gadget files are located}
}
\value{
List of class \code{gadget_main} with each name corresponding to a
type of file
and each object corresponding to a filename as used by Gadget
}
\description{
Read Gadget mainfile
}
\examples{
read_gadget_main(path = system.file(gad_mod_dir, package = "gadgetSim"))
read_gadget_main("WGTS/main.final")
}
