% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gadget_files.R
\name{read_gadget_model}
\alias{read_gadget_model}
\title{Read a Gadget model}
\usage{
read_gadget_model(main = "main", path = NULL)
}
\arguments{
\item{path}{Optional. Character vector of path to the directory where Gadget files are located}
}
\value{
A list of class \code{gadget_model}
}
\description{
This is a wrapper function for most of the \code{\link{read_gadget_*}}
functions. It reads in all of the main components for a Gadget model and
wraps them in a list of class \code{\link{gadget_model}}
}
\examples{
path <- system.file(gad_mod_dir, package = "gadgetSim")
gad_mod <- read_gadget_model(path = path)
}
