% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gadget_files.R
\name{read_gadget_stockfiles}
\alias{read_gadget_stockfiles}
\title{Read Gadget stockfiles}
\usage{
read_gadget_stockfiles(stockfiles, main = NULL, path = NULL)
}
\arguments{
\item{stockfiles}{Character vector of stocknames present in Gadget model}

\item{main}{Optional. A list of class \code{gadget_main}}

\item{path}{Optional. Character vector of path to the directory where Gadget files are located}
}
\value{
A list of class \code{gadget_stocks} consisting of
\code{gadget_stock}
one or more gadget_stock objects
}
\description{
Get stockfiles used in a Gadget model
}
\examples{
path <- system.file(gad_mod_dir, package = "gadgetSim")
main <- read_gadget_main(path = gad_mod_dir)
stocks <- read_gadget_stockfiles(main = main, path = path)
head(stocks[[1]])
}
