% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_functions.R
\name{recruit_formulas}
\alias{recruit_formulas}
\alias{simple_ssb_formula}
\alias{ricker_formula}
\alias{bev_holt_formula}
\title{Produce recruit functions in a Gadget readable formula}
\usage{
simple_ssb_formula(stockname, params = NULL)

ricker_formula(stockname, params = NULL)

bev_holt_formula(stockname, params = NULL)
}
\arguments{
\item{stockname}{A character vector of a stockname that will precede the switches for \eqn{\mu} and
\eqn{\lambda} arguments, or, optionally, a numberic vector of length 1 if SimpleSSB is desired or
of length 2 that will be used as \eqn{\mu} and \eqn{\lambda} for either the Ricker or
Beverton-Holt recruitment functions to fix those parameter values}
}
\value{
A character vector to be written to a Gadget file
}
\description{
These functions will produce recruitment curves such as Beverton-Holt or Ricker in a format that
is readable to Gadget. These are for use in creating spawnfiles.
}
\examples{
bev_holt_formula("cod")
bev_holt_formula(c(4e9, 1e8))
ricker_formula("cod")
}
