% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_functions.R
\name{recruit_funs}
\alias{recruit_funs}
\alias{simple_ssb}
\alias{ricker}
\alias{bev_holt}
\title{Basic functions for recruitment curves}
\usage{
simple_ssb(ssb, mu)

ricker(ssb, mu, lambda)

bev_holt(ssb, mu, lambda)
}
\arguments{
\item{ssb}{Numeric. The spawning stock biomass used to compute the recruitment curve}

\item{mu}{Numeric. Parameter that controls maximum number of recruits for \code{bev_holt} and
\code{ricker}}

\item{lambda}{Numeric. Controls density dependence in \code{ricker} curve and steepness in
\code{bev_holt}}
}
\value{
A numeric vector the length of \code{ssb}
}
\description{
These functions produce the basic stock-recruitment curves used by Gadget. \code{simple_ssb}
uses the equation \eqn{\mu * S}, \code{ricker} returns the number of recruits per spawning
stock biomass \eqn{S} per the Ricker formula: \deqn{R = \mu S e^{-\lambda S}}. \code{bev_holt}
returns the number of recruits using the Beverton-Holt recruitment curve
\deqn{R = (\mu * S)(\lambda + S)}
}
\examples{
curve(ricker(x, 1000, 0.1), 0, 100, xlab = "SSB", ylab = "R")
curve(bev_holt(x, 3000, 10), add = TRUE, col = "red")
}
