% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgadget_compatability.R
\name{rg_class_compatibility}
\alias{rg_class_compatibility}
\alias{gs2rg_classes}
\alias{rg2gs_classes}
\title{Convert classes of Rgadget and gadgetSim objects to be compatible with one another}
\usage{
gs2rg_classes(gs_obj)

rg2gs_classes(rg_obj)
}
\arguments{
\item{gs_obj}{An object created or read by gadgetSim (i.e. should be of class \code{gadget_*}
(i.e. \code{gadget_main}, \code{gadget_likelihood}, etc))}

\item{rg_obj}{An object created or read by Rgadget (i.e. likely will be of class \code{gadget.*}
(i.e. \code{gadget.main}, \code{gadget.printfile}, etc))}
}
\value{
The same \code{gs_obj}, but with a class appropriate for use in Rgadget
}
\description{
Much care has been taken in an attempt to keep gadgetSim and Rgadget as compatible as possible, but
Rgadget uses slightly different syntax for creating classes of objects that gadgetSim.
These functions alter classes appropriately allowing for greater compatibility between the two
packages allowing for (hopefully) seamless integration.
}
\examples{
# converting from gadgetSim to Rgadget
path <- system.file(gad_mod_dir, package = "gadgetSim")
gs_main <- read_gadget_main(path = path)
gs2rg_classes(gs_main)

# converting classes from Rgadget to gadgetSim
rg_main <- Rgadget::read.gadget.main(paste(path, "main", sep = "/"))
rg2gs_classes(rg_main)
}
