% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_functions.R
\name{selectivity}
\alias{selectivity}
\alias{logistic_selectivity}
\title{Selectivity functions}
\usage{
logistic_selectivity(length, alpha, l50, max_prop)
}
\arguments{
\item{length}{Numeric. Length to evaluate selectivity on}

\item{alpha}{Numeric. Steepness parameter}

\item{l50}{Numeric. Half-saturation parameter}

\item{max_prop}{Numeric.}
}
\value{
Numeric vector the length of \code{length} ranging from 0 to \code{max_prop}
}
\description{
Compute various selection probabilities on lengths. \code{logistic_selectivity} produces an asymptotic selection probability based on the
logistic function: \deqn{S(l) = proportion / (1 + exp(-\alpha(l - l50)))}
}
\examples{
lengths <- 1:100
logistic_selectivity(lengths, 0.2, 50, 0.1)

curve(logistic_selectivity(x, 0.2, 50, 0.1))
}
