% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simpleCap}
\alias{simpleCap}
\title{Simple function to capitalize all words of a character vector}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
A character vector with each word capitalized
}
\description{
Simple function to capitalize all words of a character vector
}
\examples{
simpleCap("may the force be with you")
}
