% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{simulate_gadget}
\alias{simulate_gadget}
\title{Create and get output from Gadget model in one fell swoop}
\usage{
simulate_gadget(..., path = NULL)
}
\value{
List of \code{data.frame}s, one for each stock, of output from
StockStdPrinter printfile component
}
\description{
This function is really just a wrapper function for \code{\link{make_gadget_model}}
and \code{\link{get_stock_std}}. It essentially just combines the functionality
of these two functions so that you can skip a line or two of writing code.
}
\examples{
# see ?make_gadget_model. The process is the same for that function, except
# here you get output
stock_stds <- simualate_gadget(time = time,
                               area = area,
                               stock = cod,
                               fleet = fleets,
                               path = gad_mod_dir)
}
