% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_}
\alias{split_}
\alias{split_tab}
\alias{split_ws}
\alias{split_list_}
\alias{split_tab_list}
\alias{split_ws_list}
\title{Split and unlist a character vector}
\usage{
split_(vec, split, ind = NULL)

split_tab(vec, ind = NULL)

split_ws(vec, ind = NULL)

split_list_(vec, split, ind = NULL)

split_tab_list(vec, ind = NULL)

split_ws_list(vec, ind = NULL)
}
\arguments{
\item{vec}{Character}

\item{split}{Character vector or object which can be coerced to such}
}
\value{
A character vector of vec split by \code{split}
}
\description{
Split and unlist a character vector
}
\examples{
split_("may the force be with you", split = " ")
split_tab("may\\tthe\\tforce\\tbe\\twith\\tyou")
}
