% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stockfile_functions.R
\name{stock_distribution_funs}
\alias{stock_distribution_funs}
\alias{normalcondfile}
\alias{normalparamfile}
\alias{numberfile}
\title{Functions to produce initial conditions or renewal distribution data}
\usage{
normalcondfile(...)

normalparamfile(...)

numberfile(...)
}
\arguments{
\item{...}{Arguments necessary for a normalcondfile, a normalparamfile, or a numberfile. See
initial conditions and renewal in the Gadget User Guide}
}
\value{
A data.frame of class normalcondfile, normalparamfile, or numberfile
}
\description{
These functions will produce a data.frame of class normalcondfile, normalparamfile, or numberfile
}
\examples{
normalparamfile(year = 1:10,
                area = 1,
                age.factor = 10,
                area.factor = 10,
                mean = vb_formula("cod", 1:10),
                stddev = 1:10,
                alpha = 0.0001,
                beta = 3)
}
