% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_functions.R
\name{suit_formulas}
\alias{suit_formulas}
\alias{make_constant_suit}
\alias{make_straightline_suit}
\alias{make_exponential_suit}
\alias{make_exponentiall50_suit}
\alias{make_richards_suit}
\alias{make_andersen_suit}
\alias{make_andersenfleet_suit}
\alias{make_gamma_suit}
\alias{make_fleet_suit}
\title{Functions to assemble suitability function lines for writing to Gadget files}
\usage{
make_constant_suit(fleet, stock, params = list("alpha"), param_names = NULL)

make_straightline_suit(fleet, stock, params = list("alpha", "beta"),
  param_names = NULL)

make_exponential_suit(fleet, stock, params = list("alpha", "beta", "gamma",
  "delta"), param_names = NULL)

make_exponentiall50_suit(fleet, stock, params = list("alpha", "l50"),
  param_names = NULL)

make_richards_suit(fleet, stock, params = as.list(paste0("p", 0:4)),
  param_names = NULL)

make_andersen_suit(fleet, stock, params = as.list(paste0("p", 0:4)),
  param_names = NULL)

make_andersenfleet_suit(fleet, stock, params = as.list(paste0("p", 0:5)),
  param_names = NULL)

make_gamma_suit(fleet, stock, params = list("alpha", "beta", "gamma"),
  param_names = NULL)

make_fleet_suit(fleet = "comm", stock = NULL, fun = "exponentiall50",
  params = NULL, param_names = NULL)
}
\arguments{
\item{fleet}{Character. The name of the fleet. Used to name switches for
suitability parameters}

\item{stock}{Character. The name of the stock to serve as the prey}

\item{params}{Optional. List of parameter names as either character vectors
or numbers. If character values are given, switches will be returned as
function parameters. If numbers are given, then the respective arguments to
the functions will be fixed.}

\item{param_names}{Optional. Character to use to name the params. This is
of use if there is more than one stock for the fleet, but the same
selectivity curve is desired}
}
\value{
A character vector of function information pasted together in a
format ready for writing to Gadget file.
}
\description{
To specify consumption for either prey or fleets Gadget requires suitability
information to be displayed in a certain manner (see Section 4.9 in
Gadget User Guide). These functions will output the correct format for each
respective suitability type. These functions are meant to be used in
conjunction when putting together Gadget stocks or fleets.
}
\examples{
make_exponential_suit("comm", "cod")
make_gamma_suit("spr", "cod")

# include multiple stocks
make_exponential_suit("comm", stock = c("immature", "mature"))
# fix a function
make_constant_suit("spr", "cod", params = 1)

# or more generally
make_fleet_suit("comm", "cod", fun = "exponentiall50",
                params = list("alpha", "l50"))
}
