% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_defaults.R
\docType{data}
\name{switches}
\alias{switches}
\alias{single_switch}
\alias{double_switch}
\alias{file_switch}
\title{Switch defaults}
\format{An object of class \code{character} of length 19.}
\usage{
single_switch

double_switch

file_switch
}
\description{
These are vectors of the switches that are used in when calling Gadget
from the command line, but you figured out a different way to implement this
You can probably delete these objects
}
\keyword{datasets}
