% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{update_model}
\alias{update_model}
\alias{update_stock}
\alias{update_fleet}
\alias{add_fleet}
\alias{rm_fleet}
\title{Update components of an object of class "gadget_model"}
\usage{
update_model(mod_obj, comp, item = NULL, ...)

update_stock(mod_obj, stockname, comp, item = NULL, ...)

update_fleet(mod_obj, fleetname, ...)

add_fleet(mod_obj, fleet)

rm_fleet(mod_obj, fleetname)
}
\arguments{
\item{mod_obj}{A list of class \code{gadget_model}, see
\code{\link{make_gadget_model}} or \code{\link{read_gadget_model}}}

\item{comp}{Character vector of the component to update (i.e. stocks, fleets,
etc.)}

\item{...}{Named arguments corresponding to the appropriate name in the
respective gadget model component (i.e. for "stocks" the named argument could
be growthparameters = c(150, 10, 1e-06, 3))}

\item{fleetname}{Character. The name of the fleet already present in a Gadget
model to either update or remove}

\item{fleet}{List of type \code{gadget_fleet} or \code{gadget_fleets} to
be added to Gadget model}
}
\value{
A list of class \code{gadget_model} the same as that of
\code{mod_obj}, but updated with the arguments given in \code{...}
}
\description{
Update components of an object of class "gadget_model"
}
\examples{
path <- system.file(gad_mod_dir, package = "gadgetSim")
mod <- read_gadget_model(path = path)
new_mod <- update_model(mod, "stocks",
                        growthparameters = c(150, 10, 1e-06, 3))
}
