% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_functions.R
\name{update_printfile}
\alias{update_printfile}
\title{Access elements of default printfiles and update them}
\usage{
update_printfile(printfile_comp, new_info)
}
\arguments{
\item{printfile_comp}{A default printfile list, see \link{printfileDefaults}}

\item{new_info}{A list of named vectors. The names correspond to the arguments to
Gadget printfile components, and the vector is the desired value}
}
\value{
A list of class corresponding to the printfile component (\code{update_printfile}
for the StockStdPrinter returns a list of class \code{stock_std}).
}
\description{
Access elements of default printfiles and update them
}
\details{
\code{update_printfile} is a generic method for the various printfile types.
Each printfile class has its own \code{update_printfile} method that returns a list of
that class. This family of functions was created to allow for the implementation of any
printfiles in functions like \code{\link{make_gadget_printfile}}. Lists used for
\code{printfile} must be one of the following classes: \code{stock_std, stock_full,
stock, predator, predator_over, prey_over, stock_prey_full, stock_prey, predator_prey,
likelihood, likelihood_summary}.
}
\examples{
cod_stock_std <- list(stockname = "cod", printfile = "cod.stock.std")
update_printfile(stock_std, cod_stock_std)
}
