% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_functions.R, R/update_printfiles.R
\name{update_printfile_dirs}
\alias{update_printfile_dirs}
\alias{update_printfile_dirs}
\title{Update path for printfile and aggfiles to include directories to write to}
\usage{
update_printfile_dirs(printfile_comp, print_dir = NULL, aggfiles = NULL)

update_printfile_dirs(printfile_comp, print_dir = NULL, aggfiles = NULL)
}
\arguments{
\item{printfile_comp}{A list of class pertaining to a Gadget printfile component,
see \code{\link{update_printfile}}}

\item{print_dir}{Character. Path of directory to print the printfiles}

\item{aggfile}{Character. Path of directory to print the aggfiles}

\item{printfile_comp}{A list of class pertaining to a Gadget printfile component,
see \code{\link{update_printfile}}}

\item{print_dir}{Character. Path of directory to print the printfiles}

\item{aggfile}{Character. Path of directory to print the aggfiles}
}
\value{
A list the same as \code{printfile}, but with updated printfile and
aggfile(s) if present

A list the same as \code{printfile}, but with updated printfile and
aggfile(s) if present
}
\description{
Update path for printfile and aggfiles to include directories to write to

Update path for printfile and aggfiles to include directories to write to
}
\examples{
cod <- list(stockname = "cod", printfile = "cod.stock.std")
cod_std <- update_printfile(stock_std, cod)
cod_std <- update_printfile_dirs(stock_std, print_dir = "printfile")
cod <- list(stockname = "cod", printfile = "cod.stock.std")
cod_std <- update_printfile(stock_std, cod)
cod_std <- update_printfile_dirs(stock_std, print_dir = "printfile")
}
