% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{us2dot}
\alias{us2dot}
\title{Replace an underscore with a period}
\usage{
us2dot(x)
}
\arguments{
\item{x}{Character vector containing an underscore}
}
\value{
Character vector with underscores replaced by period
}
\description{
Replace an underscore with a period
}
\examples{
us2dot("what's up with www_google_com")
}
