% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_functions.R, R/growth_functions.R
\name{vb_formula}
\alias{vb_formula}
\alias{vb_formula}
\title{Produce growth functions in a gadget readable formula}
\usage{
vb_formula(stockname, age)

vb_formula(stockname, age)
}
\arguments{
\item{stockname}{Character. The stockname desired for variables in the von Bertalanffy function}

\item{age}{Numeric vector of ages}

\item{stockname}{Character. The stockname desired for variables in the von Bertalanffy function}

\item{age}{Numeric vector of ages}
}
\value{
A character vector of the von Bertalanffy function the same length as \code{age}. Format is
as a Gadget formula, see \code{\link{to_gadget_formula}}

A character vector of the von Bertalanffy function the same length as \code{age}. Format is
as a Gadget formula, see \code{\link{to_gadget_formula}}
}
\description{
Produce growth functions in a gadget readable formula

Produce Gadget formula for von Bertalanffy function
}
\examples{
vb_formula("cod", 1:2)
vb_formula("cod", 1:2)
}
