% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gadget_files.R
\name{write_attr}
\alias{write_attr}
\alias{write_gadget_attributes}
\alias{write_gadget_attributes.gadget_fleet}
\alias{write_gadget_attributes.gadget_stock}
\alias{write_gadget_attributes.data.frame}
\alias{write_gadget_attributes.list}
\alias{write_gadget_attributes.gadget_spawnfile}
\alias{write_gadget_attributes.gadget_strayfile}
\alias{get_attr_filename}
\title{Functions to write attributes of Gadget components to file}
\usage{
write_gadget_attributes(gf, path = NULL, env = parent.frame())

\method{write_gadget_attributes}{gadget_fleet}(gf, path = NULL,
  env = parent.frame())

\method{write_gadget_attributes}{gadget_stock}(gf, path = NULL,
  env = parent.frame())

\method{write_gadget_attributes}{data.frame}(gf, path = NULL)

\method{write_gadget_attributes}{list}(gf, path = NULL)

\method{write_gadget_attributes}{gadget_spawnfile}(gf, path = NULL)

\method{write_gadget_attributes}{gadget_strayfile}(gf, path = NULL)

get_attr_filename(filename)
}
\arguments{
\item{gf}{The Gadget file component to be written. Should be of class
\code{gadget_fleet},
\code{gadget_stocks}, but can alternately be a \code{data.frame}}

\item{path}{Optional. Character of the path of the directory to look in}

\item{env}{An environment to search for the object named \code{path}}

\item{obj}{The attribute to be checked for filename}
}
\value{
Nothing. Writes the appropriate Gadget file, aggfile or data file to
a text file
}
\description{
These functions are used within the \code{\link{write_gadget_file}} methods
to write data from the gadget file type in R to an appropriate file in the
Gadget model. \code{write_gadget_attributes} is a generic function with
methods specific for various different types of gadget files.
}
