% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gadget_files.R
\name{write_gadget_main}
\alias{write_gadget_main}
\title{Write Gadget mainfile to a text file}
\usage{
write_gadget_main(obj = NULL, file = "main")
}
\arguments{
\item{obj}{A list also containing class \code{gadget.main}}

\item{file}{Name of the file to write to}
}
\description{
Writes a Gadget main file from an R list containing class gadget.main to a
text file to be used by Gadget
}
\examples{
write_gadget_main(gadget_main_default)
write_gadget_main()
}
