% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gadget_files.R
\name{write_gadget_model}
\alias{write_gadget_model}
\title{Write an object of class \code{gadget_model} to the respective files}
\usage{
write_gadget_model(mod_obj, path = NULL)
}
\arguments{
\item{mod_obj}{A list of class \code{gadget_model}}

\item{path}{Optional. Character of the path of the directory to write
files to}
}
\value{
Nothing. Writes the respective Gadget filetype to a text file
}
\description{
\code{write_gadget_model} takes an object of class "gadget_model" as returned
by \code{\link{read_gadget_model}}
}
\examples{
\dontrun{
path <- system.file(gad_mod_dir, package = "gadgetSim")
gad_mod <- read_gadget_model(path = path)
write_gadget_model(gad_mod, path = "different_model")
}
}
