% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{call_gadget}
\alias{call_gadget}
\title{Call Gadget executable on the command line}
\usage{
call_gadget(switches = list(s = TRUE, i = "params.in"), path = NULL,
  gadget_exe = "gadget", print_out = TRUE, print_err = TRUE)
}
\arguments{
\item{switches}{List of named switches to supply to Gadget executable.
Names correspond to the switch; values correspond to additional arguments.
If no argument is needed the value should be logical \code{TRUE}.}

\item{path}{Optional. Character vector of path to the directory where Gadget files are located}

\item{gadget_exe}{The Gadget executable to use}

\item{print_out}{Logical. Should Gadget command line output be printed.}

\item{print_err}{Logical. Should Gadget command line errors be printed.}
}
\value{
If \code{print_out = TRUE} or \code{print_err = TRUE}, a character vector giving the output
of the command, one line per string, else nothing
}
\description{
This function will call the gadget command with the given switches as arguments
}
\examples{
path <- system.file(gad_mod_dir, package = "gadgetSim")
call_gadget(switches = list(s = TRUE, i = "params.in", log = "logfile"), path = path)
call_gadget(switches = list(s = TRUE, i = "WGTS/params.final", main = "WGTS/main.final"),
            path = path)
}
