% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{def_labs}
\alias{def_labs}
\alias{aggfile_header}
\alias{gadgetfile_header}
\alias{datafile_header}
\alias{comp_lab}
\alias{comp_regex}
\title{Some default labels}
\format{An object of class \code{character} of length 1.}
\usage{
aggfile_header(af_type = "")

gadgetfile_header(gf = "")

datafile_header(df = "")

comp_lab

comp_regex
}
\arguments{
\item{af_type}{Character. The aggfile type to be included in header label (e.g. area, age, len)}

\item{gf}{Character. The type of Gadget file being written}

\item{df}{Character. The type of data file being written}
}
\value{
\code{aggfile_header} returns a character to be used as a header for printing text files
}
\description{
These are just basic default labels for some things that are used frequently
throughout functions in this package.
}
\examples{
aggfile_header("area")
}
\keyword{datasets}
