% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_likelihood.R
\name{likelihood_helper_funs}
\alias{likelihood_helper_funs}
\alias{format_lik_comp}
\alias{check_lik_data}
\alias{check_lik_names}
\alias{add_aggfile_attribute}
\alias{lik_agg_filename}
\alias{lik_data_filename}
\alias{format_datafile_attribute}
\alias{make_default_aggfile}
\alias{format_aggfile_attributes}
\title{Gadget likelihood helper functions}
\usage{
format_lik_comp(comp, ...)

check_lik_data(comp, data_cols, lik_type)

check_lik_names(comp, lik_type)

add_aggfile_attribute(agg_type, env = parent.frame())

lik_agg_filename(..., aggfile_dir = "Aggfiles")

lik_data_filename(..., datafile_dir = "Data")

format_datafile_attribute(env = parent.frame())

make_default_aggfile(comp, agg_type)

format_aggfile_attributes(comp, ...)
}
\arguments{
\item{...}{There are a number of different arguments used in these various
functions, and each depends on the function and what it is being used for.
See the source code for the \code{\link{gadget_likelihood_comps}} function
of interest to see what it does.}
}
\value{
Depends on the function
}
\description{
These functions are just helper functions that are used in the various
\code{\link{gadget_likelihood_comps}} functions. They are simply for
automation purposes to make things cleaner. These are Russian doll functions
that have helper functions within helper functions. This might make debugging
painful in the future; if this is future-you, then don't do this sort of thing
again. If this is another reader...I'm sorry.
}
