% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_printfiles.R
\name{make_aggfiles}
\alias{make_aggfiles}
\title{Check the names of a printfile component to ensure all mandatory arguments are present}
\usage{
make_aggfiles(printfile_comp)
}
\arguments{
\item{printfile_comp}{List. The printfile component to be checked}
}
\value{
NULL. Only returns an error message if any non-optional arguments are missing
}
\description{
Check the names of a printfile component to ensure all mandatory arguments are present
}
\examples{
printfile_name_check(stock_std)

cod <- update_printfile(stock_std, list(stockname = "cod", printfile = "printfile"))
printfile_name_check(cod)
}
