% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{make_gadget_printfile}
\alias{make_gadget_printfile}
\title{Values for  may only be one of the following: \code{stock_std, stock_full,
stock, predator, predator_over, prey_over, stock_prey_full, stock_prey, predator_prey,
likelihood, likelihood_summary}.}
\usage{
make_gadget_printfile(..., main = "main", file, path = NULL,
  output = "out", aggfiles = "print.aggfiles", printatstart = 1,
  steps = "all")
}
\description{
Values for  may only be one of the following: \code{stock_std, stock_full,
stock, predator, predator_over, prey_over, stock_prey_full, stock_prey, predator_prey,
likelihood, likelihood_summary}.
}
