% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gadget_files.R
\name{make_gadget_stockfile}
\alias{make_gadget_stockfile}
\title{Make Gadget stockfile}
\usage{
make_gadget_stockfile(...)
}
\arguments{
\item{...}{A list of named elements. The names of these elements must correspond to arguments
found in Gadget stockfiles (i.e. for growth, use growth = list(), for spawning, use
spawning = list(), etc), see Gadget User Guide. Chapter 4. Stock Files.}
}
\value{
A list of class \code{gadget_stock} containing information about that stock to be used
by Gadget
}
\description{
Make a list of class \code{gadget_stock} that can be used to write out a Gadget stockfile
}
\examples{
# setup basic stock information
minage <- 1
maxage <- 10
minlength <- 1
maxlength <- 100
dl <- 1
alpha <- 1e-04
beta <- 3
reflength <- seq(minlength, maxlength, dl)
stock_info <-
    list(stockname = "cod", livesonareas = 1, minage = 1, maxage = 10,
         minlength = minlength, maxlength = maxlength, dl = dl)

# setup refweightfile
stock_refwgt <-
    data.frame(len = reflength,
               weight = alpha * reflength ^ beta)

# setup growth
stock_growth <-
    list(growthfunction = "lengthvbsimple",
         growthparameters =
             c(to_gadget_formula(quote(cod.linf)), to_gadget_formula(quote(cod.k)),
               0.001, 3))
# setup naturalmortality
stock_m <- rep(0.2, 10)

# setup initial conditions
init_data <-
    normalparamfile(age = seq(minage, maxage, 1),
                    area = 1,
                    age.factor = "#cod.init.age",
                    area.factor = "#cod.init.area",
                    mean = vb_formula("cod", minage:maxage),
                    sd = 1:10,
                    alpha = alpha,
                    beta = beta)
stock_initcond <- list(normalparamfile = init_data)

# setup spawning
stock_spawnfile <-
    make_gadget_spawnfile(
      stockname = "cod",
      start_year = 1,
      end_year = 10,
      recruitment = ricker_formula("cod")
    )

# create gadget stockfile
cod <-
   make_gadget_stockfile(stock = stock_info,
                         refweightfile = stock_refwgt,
                         growth = stock_growth,
                         naturalmortality = stock_m,
                         iseaten = 1,
                         initialconditions = stock_initcond,
                         spawning = stock_spawnfile)
}
