% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_functions.R, R/update_printfiles.R
\docType{data}
\name{printfileDefaults}
\alias{printfileDefaults}
\alias{stock_std}
\alias{stock_std_args}
\alias{stock_full}
\alias{stock_full_args}
\alias{stock}
\alias{stock_args}
\alias{predator}
\alias{predator_args}
\alias{predator_over}
\alias{predator_over_args}
\alias{prey_over}
\alias{prey_over_args}
\alias{stock_prey_full}
\alias{stock_prey_full_args}
\alias{stock_prey}
\alias{stock_prey_args}
\alias{predator_prey}
\alias{predator_prey_args}
\alias{likelihood}
\alias{likelihood_args}
\alias{likelihood_summary}
\alias{likelihood_summary_args}
\alias{printfileDefaults}
\alias{stock_std}
\alias{stock_std_args}
\alias{stock_full}
\alias{stock_full_args}
\alias{stock}
\alias{stock_args}
\alias{predator}
\alias{predator_args}
\alias{predator_over}
\alias{predator_over_args}
\alias{prey_over}
\alias{prey_over_args}
\alias{stock_prey_full}
\alias{stock_prey_full_args}
\alias{stock_prey}
\alias{stock_prey_args}
\alias{predator_prey}
\alias{predator_prey_args}
\alias{likelihood}
\alias{likelihood_args}
\alias{likelihood_summary}
\alias{likelihood_summary_args}
\title{Gadget print file defaults}
\format{An object of class \code{stock_std} (inherits from \code{list}) of length 7.}
\usage{
stock_std

stock_std_args

stock_full

stock_full_args

stock

stock_args

predator

predator_args

predator_over

predator_over_args

prey_over

prey_over_args

stock_prey_full

stock_prey_full_args

stock_prey

stock_prey_args

predator_prey

predator_prey_args

likelihood

likelihood_args

likelihood_summary

likelihood_summary_args

stock_std

stock_std_args

stock_full

stock_full_args

stock

stock_args

predator

predator_args

predator_over

predator_over_args

prey_over

prey_over_args

stock_prey_full

stock_prey_full_args

stock_prey

stock_prey_args

predator_prey

predator_prey_args

likelihood

likelihood_args

likelihood_summary

likelihood_summary_args
}
\description{
Pre-baked defaults of printfile types for use in functions

Pre-baked defaults of printfile types for use in functions
}
\keyword{datasets}
