% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gadget_files.R
\name{read_gadget_fleet}
\alias{read_gadget_fleet}
\title{Read Gadget fleet file}
\usage{
read_gadget_fleet(fleetfiles, main = NULL, path = NULL)
}
\arguments{
\item{fleetfiles}{Character name of fleet file in Gadget model as specified in main file}

\item{main}{Optional. A list of class \code{gadget_main}}

\item{path}{Optional. Character vector of path to the directory where Gadget files are located}
}
\value{
A list detailing the fleets included in Gadget model,
stocks included, and suitability
}
\description{
Read Gadget fleet file
}
\examples{
path <- system.file(gad_mod_dir, package = "gadgetSim")
main <- read_gadget_main(path = path)
read_gadget_fleet(main = main, path = path)
}
