% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gadget_files.R
\name{read_gadget_stockfiles}
\alias{read_gadget_stockfiles}
\title{Read Gadget stockfiles}
\usage{
read_gadget_stockfiles(stockfiles, main = NULL, path = NULL)
}
\arguments{
\item{stockfiles}{Character vector of stocknames present in Gadget model}

\item{main}{List of class \code{gadget.main}}

\item{path}{Optional. Path to Gadget model}
}
\value{
A list of class \code{gadget.stocks} consisting of \code{gadget.stock}
one or more gadget.stock objects
}
\description{
Get stockfiles used in a Gadget model
}
\examples{
main <- read_gadget_main("main", path = system.file(gad_mod_dir, package = "gadgetSim"))
stocks <- read_gadget_stockfiles(main = main)
head(stocks[[1]])
}
