% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{strip_comments}
\alias{strip_comments}
\title{Strip comments from a character vector}
\usage{
strip_comments(x, comment = ";")
}
\arguments{
\item{x}{Character vector}
}
\value{
A character vector with comments removed
}
\description{
Strip comments from a character vector
}
\examples{
x <- c("; this is a comment in Gadget", "main\\tmainfile")
strip_comments(x)
}
