% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_functions.R
\name{suit_formulas}
\alias{suit_formulas}
\alias{constant_suit_formula}
\alias{straightline_suit_formula}
\alias{exponential_suit_formula}
\alias{exponentiall50_suit_formula}
\alias{richards_suit_formula}
\alias{andersen_suit_formula}
\alias{andersenfleet_suit_formula}
\alias{gamma_suit_formula}
\alias{fleet_suit_formula}
\title{Functions to assemble suitability function lines for writing to Gadget files}
\usage{
constant_suit_formula(fleet, stock, params = list("alpha"))

straightline_suit_formula(fleet, stock, params = list("alpha", "beta"))

exponential_suit_formula(fleet, stock, params = list("alpha", "beta", "gamma",
  "delta"))

exponentiall50_suit_formula(fleet, stock, params = list("alpha", "l50"))

richards_suit_formula(fleet, stock, params = as.list(paste0("p", 0:4)))

andersen_suit_formula(fleet, stock, params = as.list(paste0("p", 0:4)))

andersenfleet_suit_formula(fleet, stock, params = as.list(paste0("p", 0:5)))

gamma_suit_formula(fleet, stock, params = list("alpha", "beta", "gamma"))

fleet_suit_formula(fleet = "comm", stock = NULL, fun = "exponentiall50",
  params = NULL, param_names = NULL)
}
\arguments{
\item{fleet}{Character. The name of the fleet. Used in naming switches for suitability parameters}

\item{stock}{Character. The name of the stock to serve as the prey}

\item{params}{Optional. List of parameter names as either character vectors or numbers. If
character values are given, switches will be returned as function parameters. If numbers are given,
then the respective arguments to the functions will be fixed.}
}
\value{
A character vector of function information pasted together in a format ready for writing
to Gadget file.
}
\description{
To specify consumption for either prey or fleets Gadget requires suitability information to
be displayed in a certain manner (see Section 4.9 in Gadget User Guide). These functions will
output the correct format for each respective suitability type. These functions are meant to be
used in conjunction when putting together Gadget stocks or fleets.
}
\examples{
exponential_suit_formula("comm", "cod")
gamma_suit_formula("spr", "cod")

# include multiple stocks
exponential_suit_formula("comm", stock = c("immature", "mature"))
# fix a function
constant_suit_formula("spr", "cod", params = 1)

# or more generally
fleet_suit_formula("comm", "cod", fun = "exponentiall50", params = list("alpha", "l50"))
}
