% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfile_functions.R
\name{update_printfile_dirs}
\alias{update_printfile_dirs}
\title{Update path for printfile and aggfiles to include directories to write to}
\usage{
update_printfile_dirs(printfile_comp, print_dir = NULL, aggfile_dir = NULL)
}
\arguments{
\item{printfile_comp}{A list of class pertaining to a Gadget printfile component,
see \code{\link{update_printfile}}}

\item{print_dir}{Character. Path of directory to print the printfiles}

\item{aggfile_dir}{Character. Path of the directory where aggfiles will be printed}
}
\value{
A list the same as \code{printfile}, but with updated printfile and
aggfile(s) if present
}
\description{
Checks to make sure that the appropriate printfile and aggfile directories are entered in the
printfile_comp
}
\examples{
cod <- list(stockname = "cod", printfile = "cod.stock.std")
cod_std <- update_printfile(stock_std, cod)
cod_std <- update_printfile_dirs(stock_std, print_dir = "printfile")
}
