% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gadget_files.R
\name{write_aggfiles}
\alias{write_aggfiles}
\title{Write aggfiles used in a printfile_component to files in a directory}
\usage{
write_aggfiles(printfile_comp, aggfile_dir, path = NULL, fit_dir = NULL)
}
\arguments{
\item{printfile_comp}{A list of class pertaining to a Gadget printfile component,
see \code{\link{update_printfile}}}

\item{aggfile_dir}{Character. Path of the directory where aggfiles will be printed}

\item{path}{Character. Path of the Gadget model}
}
\value{
NULL. Writes aggregate files to the appropriate files in \code{aggfile_dir}
}
\description{
This function searches for aggfiles and prints them to the appropriate directory and filename
}
\examples{
path <- system.file(gad_mod_dir, package = "gadgetSim")
cod <- list(stocknames = "cod", printfile = "printfiles")
cod_stock <- update_printfile(stock, cod)
cod_stock_agg <- make_aggfiles(cod_stock, "print.aggfiles", path = path)
\dontrun{
write_aggfiles(cod_stock_agg, "print.aggfiles", path = path)
}
}
