% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gadget_files.R
\name{write_attr}
\alias{write_attr}
\alias{write_gadget_attributes}
\alias{write_gadget_attributes.gadget_fleet}
\alias{write_gadget_attributes.gadget_stock}
\alias{write_gadget_attributes.data.frame}
\alias{write_gadget_attributes.gadget_spawnfile}
\alias{get_attr_filename}
\title{Functions to write attributes of Gadget components to file}
\usage{
write_gadget_attributes(gf, path = NULL, env = parent.frame())

\method{write_gadget_attributes}{gadget_fleet}(gf, path = NULL,
  env = parent.frame())

\method{write_gadget_attributes}{gadget_stock}(gf, path = NULL,
  env = parent.frame())

\method{write_gadget_attributes}{data.frame}(df, path = NULL,
  env = parent.frame())

\method{write_gadget_attributes}{gadget_spawnfile}(gf, path = NULL,
  env = parent.frame())

get_attr_filename(obj)
}
\arguments{
\item{gf}{The Gadget file component to be written. Should be of class \code{gadget.fleet},
\code{gadget.stocks}, but can alternately be a \code{data.frame}}

\item{path}{Optional. Character of the path of the directory to look in}

\item{env}{An environment to search for the object named \code{path}}

\item{obj}{The attribute to be checked for filename}
}
\value{
Nothing. Writes the appropriate Gadget file, aggfile or data file to a text file
}
\description{
In the make_gadget_* functions it is currently implemented that files written as arguments
to various Gadget components are saved as attributes with a meta-attribute of filename
corresponding to the appropriate path given in the argument (e.g. In stockfiles, if doesspawn = 1,
there is a spawnfile argument that specifies the path to the spawning information.
\code{\link{make_gadget_stockfile}} currently saves this spawning information as an attribute
named \code{spawning}. This attribute houses the necessary spawning information and has a
meta-attribute of filename that will determine where to write the file).
}
